rearrange model files
