+++ src/kcc/datamodel/models/__init__.py
@@ -1,0 +1,2 @@
+from .core import *  # noqa
+from .klanten import *  # noqa
